+++ model/src/main/java/de/fraunhofer/iosb/ilt/faaast/service/Ensure.java
@@ -12,7 +12,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.fraunhofer.iosb.ilt.faaast.service.util;
+package de.fraunhofer.iosb.ilt.faaast.service;
 
 import java.util.function.Supplier;
